Refactor imports to align with new dir structure
