+++ components/calculator/CalcHeader.tsx
@@ -2,8 +2,8 @@ import { Button, Container, SvgIcon } from '@mui/material';
 import { useRouter } from 'next/router';
 import React from 'react';
 
-import { SharedCalcProps, StaticCalcProps } from '../../../utils/calculator.props.ts';
-import CalcStepper from '../../functional/CalcStepper.tsx';
+import { SharedCalcProps, StaticCalcProps } from '../../utils/calculator.props.ts';
+import CalcStepper from '../functional/CalcStepper.tsx';
 
 export default function CalcHeader({ page, isFirstPage }:
     { page: StaticCalcProps['page'],
